Fix build scripts

+++ build/New-CoverageReport.ps1
@@ -6,8 +6,8 @@ $ErrorActionPreference = 'Stop'
 $workspaceRoot = Split-Path $PSScriptRoot
 $projectDir = Join-Path $workspaceRoot 'tests'
 $runsettingsPath = Join-Path $projectDir 'coverlet.runsettings'
-$testProjectPath = Join-Path $projectDir 'PoshCommander.Tests.csproj'
-$testResultsDir = Join-Path $projectDir 'TestResults'
+$testProjectPath = Join-Path $projectDir 'PoshCommander.Tests.fsproj'
+$testResultsDir = Join-Path $workspaceRoot 'artifacts' 'TestResults'
 $reportOutputDir = Join-Path $testResultsDir 'CoverageReport'
 
 function Main {
@@ -32,14 +32,20 @@ function Main {
 }
 
 function GenerateCoverageResults {
-    Write-Host "Clearing results directory '$testResultsDir'" -ForegroundColor Cyan
-    Remove-Item `
-        -Force `
-        -Path "$testResultsDir\*" `
-        -Recurse
+    if (Test-Path $testResultsDir) {
+        Write-Host "Clearing results directory '$testResultsDir'" -ForegroundColor Cyan
+        Remove-Item `
+            -Force `
+            -Path "$testResultsDir\*" `
+            -Recurse
+    }
+    else {
+        Write-Host "Creating results directory '$testResultsDir'"
+    }
 
     Write-Host 'Running tests' -ForegroundColor Cyan
     dotnet test `
+        --results-directory:"$testResultsDir" `
         --settings:"$runsettingsPath" `
         "$testProjectPath" `
         | Out-Host
